+++ StarkInfra/StarkInfraTests/BrcodePreview.cs
@@ -12,23 +12,29 @@ namespace StarkInfraTests
         public readonly User user = TestUser.SetDefaultProject();
 
         [Fact]
-        public void Query()
+        public void Create()
         {
+            string payerId = "20018183000180";
             List<StaticBrcode> staticBrcodes = StaticBrcode.Query(limit: 3).ToList();
             List<DynamicBrcode> dynamicBrcodes = DynamicBrcode.Query(limit: 3).ToList();
+
             List<BrcodePreview> brcodes = new List<BrcodePreview>
             {
                 new BrcodePreview(
-                        id: staticBrcodes[0].ID
+                        id: staticBrcodes[0].ID,
+                        payerId: payerId
                     ),
                 new BrcodePreview(
-                        id: staticBrcodes[1].ID
+                        id: staticBrcodes[1].ID,
+                        payerId: payerId
                     ),
                 new BrcodePreview(
-                        id: dynamicBrcodes[0].ID
+                        id: dynamicBrcodes[0].ID,
+                        payerId: payerId
                     ),
                 new BrcodePreview(
-                        id: dynamicBrcodes[1].ID
+                        id: dynamicBrcodes[1].ID,
+                        payerId: payerId
                     )
             };
 
@@ -41,6 +47,7 @@ namespace StarkInfraTests
                 TestUtils.Log(preview);
                 Assert.NotNull(preview.ID);
                 Assert.Equal(preview.ID, brcodes[index].ID);
+                Assert.Equal(preview.PayerId, payerId);
                 index++;
             }
         }